cut/requirements.txt
aioconsole==0.6.1
aiofiles==22.1.0
aiogTTS==1.1.1
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.18.0
altgraph==0.17.3
anyio==3.6.2
appdirs==1.4.4
arrow==1.3.0
art==6.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.27.0
attrs==22.2.0
auto-py-to-exe==2.36.0
beautifulsoup4==4.12.2
bottle==0.12.25
bottle-websocket==0.2.9
braceexpand==0.1.7
Brotli==1.0.9
bs4==0.0.1
button-paginator==1.0.1
cachetools==5.3.0
cairocffi==1.6.1
CairoSVG==2.7.1
cashews==6.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
colored==1.4.4
colorgram.py==1.2.0
commonmark==0.9.1
config==0.5.1
cssselect2==0.7.0
dacite==1.8.0
dataclass-factory==2.16
datadispatch==1.0.0
DateTime==5.2
deep-translator==1.11.4
defusedxml==0.7.1
discord==1.7.3
discord-ios==1.1.1
discord-paginator==0.3.1
discord-protos==0.0.2
discord-py-interactions==5.10.0
discord-py-slash-command==3.0.3
discord-typings==0.7.0
discord.py @ git+https://github.com/rapptz/discord.py@a5d03d4a1e6f4314518f6f55d8b7130ab3f4336e
discordlevelingcard==0.6.0
discum==1.4.1
dnspython==2.3.0
docopt==0.6.2
easygui==0.98.3
Eel==0.16.0
emoji==2.2.0
exceptiongroup==1.1.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filetype==1.2.0
Flag==0.1.1
Flask==2.2.3
fortnite-api==2.5.4
frozenlist==1.4.0
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
google-api-core==2.11.0
google-auth==2.17.2
google-cloud-dialogflow==2.21.0
googleapis-common-protos==1.59.0
gratient==0.4
greenlet==2.0.2
grpcio==1.53.0
grpcio-status==1.53.0
gTTS==2.3.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
humanfriendly==10.0
humanize==4.8.0
idna==3.4
import-expression==1.1.4
importlib-metadata==6.8.0
iniconfig==2.0.0
instaloader==4.10.1
intents==0.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
jishaku==2.5.1
joblib==1.2.0
Js2Py==0.74
keyboard==0.13.5
line-profiler==4.0.3
MarkupSafe==2.1.2
modAL==0.4.1
MouseInfo==0.1.3
multidict==6.0.4
munch==4.0.0
mutagen==1.47.0
numpy==1.24.2
orjson==3.9.9
outcome==1.2.0
packaging==23.0
Paginator==0.5.1
pandas==2.0.0
pefile==2023.2.7
pexpect==4.8.0
pg==0.1
PIL-Tools==1.1.0
Pillow==9.5.0
pipwin==0.5.2
pluggy==1.3.0
pomice==2.7.2
postgres==4.0
proto-plus==1.22.2
protobuf==4.22.1
proxy==0.0.1
psutil==5.9.4
psycopg2-binary==2.9.9
psycopg2-pool==1.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pydantic==1.10.13
pydub==0.25.1
pyee==8.2.2
pyfiglet==0.8.post1
PyGetWindow==0.0.9
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
pyjsparser==2.7.1
pylibcheck==0.2.2
pymongo==4.3.3
PyMsgBox==1.0.9
PyNaCl==1.5.0
PyOpenGL==3.1.6
pyparsing==3.1.0
pyperclip==1.8.2
pypiwin32==223
pyppeteer==1.0.2
pypresence==4.2.1
PyPrind==2.11.3
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==0.1.28
pySmartDL==1.3.4
PySocks==1.7.1
pystyle==2.9
pytesseract==0.3.10
pytest==7.4.2
pytest-asyncio==0.20.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytweening==1.0.4
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==5.4.1
redis==5.0.1
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.8.3
selfcord.py @ git+https://github.com/dolfies/discord.py-self@c333695b6798ecc53aa8edd9b6a9b9136b3089e8
shazamio==0.4.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==1.4.49
Strip==0.0.1.dev14
tabulate==0.8.10
tasks==2.8.0
tasksio==0.0.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tk==0.1.0
tls-client==0.2.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
ua-parser==0.16.1
ui==0.1.4
urllib3==1.26.15
user-agent==0.1.10
uwuipy==0.1.7
webdriver-manager==3.8.5
webencodings==0.5.1
websocket==0.2.1
websocket-client==0.59.0
websockets==10.4
Werkzeug==2.2.3
whichcraft==0.6.1
wsproto==1.2.0
xxhash==3.4.1
yarl==1.8.2
youtube-dl==2021.12.17
yt-dlp==2023.10.13
zipp==3.17.0
zope.event==4.6
zope.interface==6.0
